Bugfix mhs calibration
